Edited Enemy Patrol code
Done:
Each enemy in the level creates its own instance of the enemy states to prevent a bug where all enemies use the same patrol point distances.

Changed the speed of the patrol lerp so that it now factors in the distance the points are apart from each other so that each enemy's movement speed is consistent with the 'walkspeed' variable.